Move API logic into useInstructors hook
